⚡️ Speed up method `PromptTemplate.get_identifiers` by 86% [codeflash] (#79)
* ⚡️ Speed up method `PromptTemplate.get_identifiers` by 86%
To optimize the `get_identifiers` method in the `PromptTemplate` class, we can make a few potential improvements. However, it's worth noting that this particular method appears straightforward and already quite efficient. In this case, there might be only subtle changes we can make to optimize it further, particularly if the set operations are already handled efficiently in Python. We'll ensure that it's as optimized as possible.



Explanation.
- The method `self._args.keys()` returns a view on the dictionary’s keys, and converting these keys directly into a set is efficient and straightforward.
- We can directly use `set(self._args)` instead of converting `self._args.keys()` to a list and then to a set.

As the original method was already quite performant, this adjustment ensures that we are leveraging Python’s efficient internal handling of dictionary keys and set creation.

* remove comment

---------

Co-authored-by: codeflash-ai[bot] <148906541+codeflash-ai[bot]@users.noreply.github.com>